beginning boilerplate and settings
