+++ package.json
@@ -14,12 +14,16 @@
   "devDependencies": {
     "@sveltejs/adapter-auto": "next",
     "@sveltejs/kit": "next",
+    "@tailwindcss/typography": "^0.5.2",
     "@types/cookie": "^0.4.1",
+    "autoprefixer": "^10.4.2",
+    "postcss": "^8.4.6",
     "prettier": "^2.4.1",
     "prettier-plugin-svelte": "^2.4.0",
     "svelte": "^3.44.0",
     "svelte-check": "^2.2.6",
     "svelte-preprocess": "^4.9.4",
+    "tailwindcss": "^3.0.23",
     "tslib": "^2.3.1",
     "typescript": "^4.4.3"
   },
@@ -29,6 +33,7 @@
     "@lukeed/uuid": "^2.0.0",
     "axios": "^0.24.0",
     "cookie": "^0.4.1",
+    "daisyui": "^2.1.0",
     "esbuild-wasm": "^0.14.9",
     "localforage": "^1.10.0",
     "monaco-editor": "^0.31.1"
